+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changelog
 -------------------
 
 - Added void output formats for each domain which discard the conversion results.
+- Added option to LocalFilenameSource which tells it to write the files it reads to a list-file.
 
 0.1.1 (2021-05-20)
 -------------------